requirements.txt
imageio==2.34.0
numpy==1.26.4
pillow==10.2.0
python-magic==0.4.27
tifffile==2024.2.12
